--- common/src/main/java/com/github/camotoy/geyserblockplatform/common/Permissions.java
@@ -1,5 +1,0 @@
-package com.github.camotoy.geyserblockplatform.common;
-
-public final class Permissions {
-    public static final String BYPASS = "geyserblockplatform.bypass";
-}
