--- common/src/main/java/com/github/camotoy/geyserblockplatform/common/config/ScalarParent.java
@@ -1,9 +1,0 @@
-package com.github.camotoy.geyserblockplatform.common.config;
-
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ScalarParent {
-    String mapKey();
-}
