--- common/src/main/java/com/github/camotoy/geyserblockplatform/common/platformchecker/BedrockPlatformChecker.java
@@ -1,10 +1,0 @@
-package com.github.camotoy.geyserblockplatform.common.platformchecker;
-
-import com.github.camotoy.geyserblockplatform.common.Profile;
-
-import java.util.UUID;
-
-public interface BedrockPlatformChecker {
-    boolean isBedrockPlayer(UUID uuid);
-    Profile profile(UUID uuid);
-}
