+++ common/src/main/java/dev/kejona/geyserblockplatform/common/Permissions.java
@@ -1,0 +1,5 @@
+package dev.kejona.geyserblockplatform.common;
+
+public final class Permissions {
+    public static final String BYPASS = "geyserblockplatform.bypass";
+}
